render filter and struct separation * Users now able to filter which tables to render when loading previously saved check
